refactor leaderboard logic into shared utilities
